+++ NEWS.md
@@ -9,6 +9,7 @@ Version 0.7.3
  - fix: syntax issues not correctly cleared when the file was saved as another name.
  - enhancement: when running a program, redirect a data file to its stdin
  - fix: can't correctly handle '&&' and '||' in the #if directive (and correctly parse windows.h header file)
+ - fix: crash when create an empty project
 
 Version 0.7.2
  - fix: rainbow parenthesis stop functioning when change editor's general options